+++ apps/web-giddh/src/app/app.module.ts
@@ -38,20 +38,18 @@ import { ScrollingModule } from '@angular/cdk/scrolling';
 import { MatButtonModule } from '@angular/material/button';
 import { FormFieldsModule } from './theme/form-fields/form-fields.module';
 import { VerifySubscriptionTransferOwnershipModule } from './verify-subscription-transfer-ownership/verify-subscription-transfer-ownership.module';
-import { EnvironmentService } from './services/enviroment.service';
+import { ReplaceEnvGuard } from './decorators/replace-env-variables-guard';
 
 // Application wide providers
 const APP_PROVIDERS = [
     ...APP_RESOLVER_PROVIDERS,
-    { provide: APP_BASE_HREF, useValue: IS_ELECTRON_WA ? './' : AppUrl + APP_FOLDER }
+    { provide: APP_BASE_HREF, useValue: IS_ELECTRON_WA ? './' : AppUrl + APP_FOLDER }, ReplaceEnvGuard
 ];
 
 // tslint:disable-next-line:prefer-const
 let CONDITIONAL_IMPORTS = [];
 
 export function localStorageSyncReducer(reducer: ActionReducer<any>): ActionReducer<any> {
-    console.log(reducer);
-
     return localStorageSync({ keys: ['session', 'permission', 'branchConsolidated'], rehydrate: true, storage: localStorage })(reducer);
 }
 
@@ -76,6 +74,27 @@ if (giddhRegion === "UK") {
     localStorage.setItem("Country-Region", "GL");
 }
 
+// Function to create service config
+function createServiceConfig() {
+    let whiteLabel = document.cookie
+        .split('; ')
+        .find(cookie => cookie.startsWith('whiteLabel='))
+        ?.split('=')[1];
+    console.log('white label',whiteLabel);
+
+    const config = whiteLabel;
+    // return {
+    //     apiUrl: config?.giddhWhiteLabel?.apiDomainName
+    //         ? `https://${config.giddhWhiteLabel.apiDomainName}/`
+    //         : Configuration.ApiUrl,
+    //     appUrl: config?.giddhWhiteLabel?.domainName
+    //         ? `https://${config.giddhWhiteLabel.domainName}/`
+    //         : Configuration.AppUrl
+    // };
+}
+
+
+
 /**
  * `AppModule` is the main entry point into Angular2's bootstraping process
  */
@@ -91,6 +110,11 @@ if (giddhRegion === "UK") {
      */
     imports: [
         BrowserModule,
+        RouterModule.forRoot(ROUTES, {
+            useHash: IS_ELECTRON_WA,
+            onSameUrlNavigation: 'reload',
+            preloadingStrategy: QuicklinkStrategy
+        }),
         BrowserAnimationsModule,
         FormsModule,
         ReactiveFormsModule,
@@ -105,11 +129,6 @@ if (giddhRegion === "UK") {
         ToastrModule.forRoot({ preventDuplicates: true, maxOpened: 3 }),
         StoreModule.forRoot(reducers, { metaReducers, runtimeChecks: { strictStateImmutability: false, strictActionImmutability: false } }),
         ScrollingModule,
-        RouterModule.forRoot(ROUTES, {
-            useHash: IS_ELECTRON_WA,
-            onSameUrlNavigation: 'reload',
-            preloadingStrategy: QuicklinkStrategy
-        }),
         QuicklinkModule,
         MatSnackBarModule,
         SnackBarModule,
@@ -124,10 +143,12 @@ if (giddhRegion === "UK") {
      * enableTracing: true,
      */
     providers: [
+        environment.ENV_PROVIDERS,
+        APP_PROVIDERS,
+        WindowRef,
         {
             provide: ServiceConfig,
-            useFactory: (environmentService: EnvironmentService) => environmentService.createServiceConfig(),
-            deps: [EnvironmentService], // Inject the service
+            useValue: createServiceConfig()
         },
         {
             provide: HTTP_INTERCEPTORS,
@@ -137,6 +158,7 @@ if (giddhRegion === "UK") {
             provide: ErrorHandler,
             useClass: ExceptionLogService
         },
+        ReplaceEnvGuard,
         CustomPreloadingStrategy
     ]
 })
